+++ .gitignore
@@ -1,0 +1,4 @@
+pyvenv.cfg
+Include
+Lib
+Scripts
